requirements.txt
flask
flask-cors
tensorflow
scikit-learn
pandas
numpy
yfinance
requests
python-dotenv
finnhub-python
gunicorn
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
